chore: move stray css files from src to src/styles
