+++ packages/server-admin-ui/src/styles/blinking-circle.css
