+++ packages/server-admin-ui/src/styles/fa-pulse.css
